+++ DelvUI/changelog.md
@@ -6,6 +6,9 @@
     + There's a new section under Customization > Bar Textures where custom textures can be added.
     + Textures can be previewed in Customization > Bar Textures.
     + Also in this section it is possible to apply a particular Bar Texture and Draw Mode to all bars with one click.
+
+- Added settings for Enemy Nameplate Castbars to match the width or height of the Health Bar:
+    + These are intended to be used when the Health Bar size changes for targeted units.
     
 - Moved "Experience Bar", "GCD Indicator", "Pull Timer", "Limit Break" and "MP Ticker" from "Misc" to the new "Other Elements" section.
 - Moved "Fonts" from "Misc" to the new "Customization" section.